+++ src/main/java/vendingmachine/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine.machine;
+package vendingmachine;
 
 public enum Coin {
     COIN_500(500),
